client/examples/inventory/requirements.txt
asgiref==3.8.1
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
blinker==1.8.1
cffi==1.16.0
channels==4.1.0
click==8.1.7
constantly==23.10.4
cryptography==42.0.5
daphne==4.1.2
Django==5.0.6
grpcio==1.64.1
grpcio-tools==1.64.1
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
Jinja2==3.1.3
MarkupSafe==2.1.5
protobuf==5.27.1
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.8.0
pyOpenSSL==24.1.0
service-identity==24.1.0
six==1.16.0
sqlparse==0.5.0
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.12.2
zope.interface==6.4.post2
